+++ themes/WebGuitarPro_AngularJs/assets/js/pages/profile.js
@@ -1,9 +1,10 @@
-require( '../../css/profile.css' );
-
-import { VsDisplayPassword } from '../includes/password-generator.js';
-import { VsPath } from '../includes/fos_js_routes.js';
-import { VsTranslator, VsLoadTranslations } from '../includes/bazinga_js_translations.js';
+import { VsDisplayPassword } from '@/js/includes/password-generator.js';
+import { VsPath } from '@/js/includes/fos_js_routes.js';
+import { VsTranslator, VsLoadTranslations } from '@/js/includes/bazinga_js_translations.js';
 VsLoadTranslations(['VSApplicationBundle']);
+
+require( '@/js/includes/bootstrap-5/file-input.js' );
+require( '../../css/profile.css' );
 
 $( function()
 {